CollabWorkDW/src/main/java/com/sahajjain/api/Saying.java
package com.sahajjain.api;

import com.fasterxml.jackson.annotation.JsonProperty;
import org.hibernate.validator.constraints.Length;

import javax.annotation.security.RolesAllowed;

public class Saying {
    private long id;
    private String content;

    public Saying() {
    }

    public Saying(long id, String content) {
        this.id = id;
        this.content = content;
    }

    @JsonProperty
    public long getId() {
        return id;
    }

    @JsonProperty
    public String getContent() {
        return content;
    }
}